repo cleanup, back functionality dashboard
