--- lib/models/deals/Deal.ts
@@ -1,29 +1,0 @@
-import { DealMetadata, DealType, Party } from "lib/types";
-
-export class Deal {
-
-    id: string = null;
-
-    type: DealType = null;
-
-    primaryParty: Party = null;
-
-    secondaryParty: Party  = null;
-
-    cosigners?: Party[] = undefined;
-
-    metadata: DealMetadata  = null;
-
-    specialClause?: string = null;
-
-
-    constructor() {
-        if (this.constructor == Deal) {
-            throw new Error('Deal is an abstract class. Please use one of the DealType constructors')
-        }
-
-    }
-    
-}
-
-
