--- lib/models/deals/index.ts
@@ -1,3 +1,0 @@
-export * from './IOU'
-export * from './Deal'
-export * from './schema'
