+++ lib/models/index.ts
@@ -1,4 +1,3 @@
-export * from "./deals";
 export * from "./Folder";
 
 export * from "./Kernel";
